requirements.txt
django
djangorestframework
setuptools
twine
pycryptodome
python-dotenv